+++ R/data_chester.R
@@ -23,9 +23,10 @@
 #' "Higher Rates Of Hate Crimes Are Tied To Income Inequality"
 #' \url{https://fivethirtyeight.com/features/higher-rates-of-hate-crimes-are-tied-to-income-inequality/}.
 #'
-#' @format A data frame with 51 rows representing US states and DC and 12 variables:
+#' @format A data frame with 51 rows representing US states and DC and 13 variables:
 #' \describe{
 #'   \item{state}{State name}
+#'   \item{state_abbrev}{State abbreviation}
 #'   \item{median_house_inc}{Median household income, 2016}
 #'   \item{share_unemp_seas}{Share of the population that is unemployed (seasonally adjusted), Sept. 2016}
 #'   \item{share_pop_metro}{Share of the population that lives in metropolitan areas, 2015}